fix(deploy): resolve 440 whatsapp socket conflict to allow successful Render deployments, configure 1s real-time tracking, fix Samsung Pass autofill, and improve Android 11+ background location redirection

+++ www/js/modules/radar-v126.js
@@ -158,7 +158,7 @@ const RadarModule = (() => {
             </button>
             <div class="radar-legend" id="radarLegend">
                 <span class="radar-legend-item">🚗 Choferes activos: <strong id="radarActiveCount">0</strong></span>
-                <span class="radar-legend-item">🕐 Actualización: <strong>Tiempo real (2s)</strong></span>
+                <span class="radar-legend-item">🕐 Actualización: <strong>Tiempo real (1s)</strong></span>
             </div>
         `;
